+++ src/App.js
@@ -3,7 +3,6 @@ import { Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
 import Category from "./pages/Category";
 import Product from "./pages/ProductPage";
-
 const App = () => {
   return (
     <Routes>
