+++ src/components/ProductTable.jsx
@@ -1,0 +1,45 @@
+import React, { useState } from "react";
+// import Product from "../components/Product";
+// import Product from "../components/ProductCard";
+import { styled } from "@mui/material/styles";
+import TableCell, { tableCellClasses } from "@mui/material/TableCell";
+import TableRow from "@mui/material/TableRow";
+import Button from "@mui/material/Button";
+import "../App.css";
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.common.black,
+    color: theme.palette.common.white,
+  },
+  [`&.${tableCellClasses.body}`]: {
+    fontSize: 14,
+  },
+}));
+
+const StyledTableRow = styled(TableRow)(({ theme }) => ({
+  "&:nth-of-type(odd)": {
+    backgroundColor: theme.palette.action.hover,
+  },
+  // hide last border
+  "&:last-child td, &:last-child th": {
+    border: 0,
+  },
+}));
+
+export default function CustomizedTables(props) {
+  return (
+    <StyledTableRow key={props.product.title}>
+      <StyledTableCell component="th" scope="row">
+        {props.product.title}
+      </StyledTableCell>
+      <StyledTableCell align="right">{props.product.type}</StyledTableCell>
+      <StyledTableCell align="right">{props.product.price}$</StyledTableCell>
+
+      <StyledTableCell align="right">
+        <Button variant="contained" color="success">
+          click
+        </Button>
+      </StyledTableCell>
+    </StyledTableRow>
+  );
+}
